extracted clearFunction into separate file
